fix: reminders
